+++ UnitTests/Sources/TimelineMediaPreviewViewModelTests.swift
@@ -182,8 +182,7 @@ struct TimelineMediaPreviewViewModelTests {
         // When confirming the redaction.
         let deferred = deferFulfillment(viewModel.actions) { $0 == .dismiss }
         
-        // The redaction runs in an unstructured task that races against the dismiss action,
-        // so wait for the mock to be called instead of asserting right after the dismiss.
+        // The redaction runs in an unstructured task, so wait for the call rather than asserting after the dismiss.
         await waitForConfirmation { confirmation in
             timelineController.redactClosure = { _ in confirmation() }
             context.send(viewAction: .redactConfirmation(item: mediaItem))
@@ -191,7 +190,6 @@ struct TimelineMediaPreviewViewModelTests {
         
         // Then the item should be redacted and the view should be dismissed.
         try await deferred.fulfill()
-        #expect(timelineController.redactCalled)
     }
     
     @Test
